refactor: extract shared alert rules and address review feedback
- Extract containerVolumeAlertRules() helper to eliminate duplication
  between Zalando and CNPG PrometheusRules (review comment #3)
- Fix pod name regex [0-9] → [0-9]+ to match instances >= 10 (#4)
- Split CNPG PrometheusRule tests into dedicated file (#5)
- Review comments #1 and #2 (golden test config) were already fixed
  in previous commit
